refactor: extract prompt pipeline helpers
