+++ tests/common/prompting/promptPipelineTestBed.js
@@ -7,6 +7,11 @@
 import { AIPromptPipeline } from '../../../src/prompting/AIPromptPipeline.js';
 import { expect } from '@jest/globals';
 import {
+  AIPromptPipelineDependencySpec,
+  expectSuccessfulGeneration,
+  expectGenerationFailure,
+} from './pipelineHelpers.js';
+import {
   createMockLogger,
   createMockLLMAdapter,
   createMockAIGameStateProvider,
@@ -36,38 +41,6 @@ const PipelineFactoryMixin = createServiceFactoryMixin(
     }),
   'pipeline'
 );
-
-/**
- * @typedef {object} DependencySpecEntry
- * @property {RegExp} error - Expected error when dependency is missing.
- * @property {string[]} methods - Methods required on the dependency.
- */
-
-/**
- * @description Defines how {@link AIPromptPipeline} constructor dependencies
- * should be validated within tests. Each property specifies the expected error
- * regex and required method names for that dependency.
- * @type {Object<string, DependencySpecEntry>}
- */
-export const AIPromptPipelineDependencySpec = {
-  llmAdapter: {
-    error: /ILLMAdapter/,
-    methods: ['getAIDecision', 'getCurrentActiveLlmId'],
-  },
-  gameStateProvider: {
-    error: /IAIGameStateProvider/,
-    methods: ['buildGameState'],
-  },
-  promptContentProvider: {
-    error: /IAIPromptContentProvider/,
-    methods: ['getPromptData'],
-  },
-  promptBuilder: {
-    error: /IPromptBuilder/,
-    methods: ['build'],
-  },
-  logger: { error: /ILogger/, methods: ['info'] },
-};
 
 /**
  * @description Utility class for unit tests that need an AIPromptPipeline with common mocks.
@@ -162,70 +135,6 @@ export class AIPromptPipelineTestBed extends PipelineFactoryMixin(
     );
     this.mocks.promptBuilder.build.mockResolvedValue(finalPrompt);
   }
-
-  /**
-   * Generates a prompt and verifies that all mocked dependencies were called
-   * with the values configured by setupMockSuccess.
-   *
-   * @param {object} params - Options for generation and assertions.
-   * @param {import('../../../src/entities/entity.js').default} params.actor - Actor for the prompt.
-   * @param {import('../../../src/turns/interfaces/ITurnContext.js').ITurnContext} params.context - Turn context.
-   * @param {import('../../../src/turns/dtos/actionComposite.js').ActionComposite[]} params.actions - Available actions.
-   * @param {string} params.expectedPrompt - Expected final prompt string.
-   * @param {string} [params.llmId] - Optional LLM ID override.
-   * @param {object} [params.gameState] - Optional game state override.
-   * @param {object} [params.promptData] - Optional prompt data override.
-   * @param {string} [params.finalPrompt] - Optional final prompt override.
-   * @returns {Promise<void>} Resolves when assertions pass.
-   */
-  async expectSuccessfulGeneration({
-    actor,
-    context,
-    actions,
-    expectedPrompt,
-    llmId,
-    gameState,
-    promptData,
-    finalPrompt,
-  }) {
-    this.setupMockSuccess({ llmId, gameState, promptData, finalPrompt });
-    const prompt = await this.generate(actor, context, actions);
-    expect(prompt).toBe(expectedPrompt);
-
-    const {
-      llmId: _llmId,
-      gameState: _gameState,
-      promptData: _promptData,
-    } = this._successOptions;
-    expect(this.llmAdapter.getCurrentActiveLlmId).toHaveBeenCalledTimes(1);
-    expect(this.gameStateProvider.buildGameState).toHaveBeenCalledWith(
-      actor,
-      context,
-      this.logger
-    );
-    expect(this.promptContentProvider.getPromptData).toHaveBeenCalledWith(
-      expect.objectContaining({ ..._gameState, availableActions: actions }),
-      this.logger
-    );
-    expect(this.promptBuilder.build).toHaveBeenCalledWith(_llmId, _promptData);
-  }
-
-  /**
-   * Applies a mutation to the test bed's mocks and expects generation to fail.
-   *
-   * @description Applies a mutation to the mocks and verifies that
-   *   {@link AIPromptPipelineTestBed#generateDefault} rejects with the given
-   *   error.
-   * @param {(bed: this) => void} mutateFn - Function that mutates the test
-   *   bed's mocks before generation.
-   * @param {string|RegExp|Error} expectedError - Error expected from the
-   *   generation call.
-   * @returns {Promise<void>} Resolves when the assertion completes.
-   */
-  async expectGenerationFailure(mutateFn, expectedError) {
-    mutateFn(this);
-    await expect(this.generateDefault()).rejects.toThrow(expectedError);
-  }
 }
 
 /**
@@ -247,4 +156,9 @@ export const {
   },
 });
 
+export {
+  AIPromptPipelineDependencySpec,
+  expectSuccessfulGeneration,
+  expectGenerationFailure,
+};
 export default AIPromptPipelineTestBed;
